+++ src/Filters/src/ErrorMapper.php
@@ -13,6 +13,10 @@ final class ErrorMapper
     ) {
     }
 
+    /**
+     * @param array<string, string> $errors
+     * @return array<string, string>
+     */
     public function mapErrors(array $errors): array
     {
         // De-mapping